# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Trade Finder API
  version: "2.0"
  description: |
    OpenAPI specification for the AI Trade Finder project.

    Notes:
    - Authentication is not implemented (all endpoints are open).
    - Some controllers return empty bodies on error (e.g., 500 with no JSON).
    - Several fields are intentionally flexible objects (Map<String,Object>); those are modeled as `type: object` with `additionalProperties: true`.
servers:
  - url: http://localhost:8082
    description: Local default
...
paths:
...
  /api/v1/market-events:
    post:
      tags: [Market Events v1]
      summary: Create core market event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoreMarketEvent'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoreMarketEvent'
        '500':
          description: Internal error (often empty body)
    get:
      tags: [Market Events v1]
      summary: List all market events
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CoreMarketEvent'
        '500':
          description: Internal error (often empty body)
  /api/v1/market-events/{id}:
    get:
      tags: [Market Events v1]
      summary: Get market event by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoreMarketEvent'
        '404':
          description: Not found
        '500':
          description: Internal error (often empty body)
    put:
      tags: [Market Events v1]
      summary: Update market event by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoreMarketEvent'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoreMarketEvent'
        '404':
          description: Not found
        '500':
          description: Internal error (often empty body)
    delete:
      tags: [Market Events v1]
      summary: Delete market event by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
        '500':
          description: Internal error (often empty body)
  /api/v1/market-events/symbol/{symbol}:
    get:
      tags: [Market Events v1]
      summary: Query events by symbol and time range
      parameters:
        - in: path
          name: symbol
          required: true
          schema: { type: string }
        - in: query
          name: start
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: end
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CoreMarketEvent'
        '500':
          description: Internal error (often empty body)
  /api/v1/market-events/symbol/{symbol}/timeframe/{timeframe}:
    get:
      tags: [Market Events v1]
      summary: Query events by symbol, timeframe, and time range
      parameters:
        - in: path
          name: symbol
          required: true
          schema: { type: string }
        - in: path
          name: timeframe
          required: true
          schema: { type: string }
        - in: query
          name: start
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: end
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CoreMarketEvent'
        '500':
          description: Internal error (often empty body)
...
  /api/v1/conversations:
    post:
      tags: [Conversations v1]
      summary: Create a conversation session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
            examples:
              trade_analysis:
                value:
                  conversationType: TRADE_ANALYSIS
                  symbol: NQ
                  userId: user123
                  expiryHours: 24
      responses:
        '200':
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIConversation'
  /api/v1/conversations/{conversationId}:
    get:
      tags: [Conversations v1]
      summary: Get conversation by ID
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIConversation'
        '404':
          description: Not found
  /api/v1/conversations/{conversationId}/stats:
    get:
      tags: [Conversations v1]
      summary: Get conversation stats
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Stats map
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: Not found
  /api/v1/conversations/symbol/{symbol}:
    get:
      tags: [Conversations v1]
      summary: Get active conversations for a symbol
      parameters:
        - in: path
          name: symbol
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AIConversation'
  /api/v1/conversations/{conversationId}/complete:
    post:
      tags: [Conversations v1]
      summary: Mark conversation completed
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
components:
  schemas:
...
    CoreMarketEvent:
      type: object
      properties:
        id: { type: string, nullable: true }
        symbol: { type: string }
        timeframe: { type: string }
        indicatorName: { type: string }
        rawMessage: { type: string, nullable: true }
        ingestedTs: { type: string, format: date-time, nullable: true }
        meta:
          type: object
          nullable: true
          additionalProperties: true
        queued: { type: boolean }
        transformAttempts: { type: integer, format: int32 }
    CreateConversationRequest:
      type: object
      properties:
        conversationType:
          type: string
          description: TRADE_FOLLOWUP | TRADE_ANALYSIS | MARKET_ANALYSIS | WORKFLOW (stored as string)
        symbol: { type: string, nullable: true }
        userId: { type: string, nullable: true }
        expiryHours: { type: integer, format: int32, nullable: true }
    AIConversation:
      type: object
      properties:
        id: { type: string, nullable: true }
        conversationId: { type: string }
        symbol: { type: string, nullable: true }
        conversationType: { type: string, nullable: true }
        userId: { type: string, nullable: true }
        tradeId: { type: string, nullable: true }
        entityType: { type: string, nullable: true }
        entityId: { type: string, nullable: true }
        entitySnapshot:
          type: object
          nullable: true
          additionalProperties: true
        turns:
          type: array
          items:
            $ref: '#/components/schemas/ConversationTurn'
        status:
          type: string
          enum: [ACTIVE, COMPLETED, EXPIRED]
        createdAt: { type: string, format: date-time, nullable: true }
        lastActivityAt: { type: string, format: date-time, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        contextData:
          type: object
          nullable: true
          additionalProperties: true
        userPreferences:
          type: object
          nullable: true
          additionalProperties: true
        tags:
          type: array
          items: { type: string }
          nullable: true
    ConversationTurn:
      type: object
      properties:
        responseId: { type: string, nullable: true }
        requestId: { type: string, nullable: true }
        userInputSummary: { type: string, nullable: true }
        aiOutputSummary: { type: string, nullable: true }
        model: { type: string, nullable: true }
        timestamp: { type: string, format: date-time, nullable: true }
        tokensUsed: { type: integer, format: int32, nullable: true }
        metadata:
          type: object
          nullable: true
          additionalProperties: true